Move system_transaction out of src/
