+++ benches/banking_stage.rs
@@ -13,10 +13,10 @@ use solana::banking_stage::{BankingStage, NUM_THREADS};
 use solana::entry::Entry;
 use solana::mint::Mint;
 use solana::packet::to_packets_chunked;
-use solana::system_transaction::SystemTransaction;
 use solana_sdk::hash::hash;
 use solana_sdk::pubkey::Pubkey;
 use solana_sdk::signature::{KeypairUtil, Signature};
+use solana_sdk::system_transaction::SystemTransaction;
 use solana_sdk::transaction::Transaction;
 use std::iter;
 use std::sync::mpsc::{channel, Receiver};
