+++ sdk/src/lib.rs
@@ -16,6 +16,7 @@ pub mod signature;
 pub mod storage_program;
 pub mod system_instruction;
 pub mod system_program;
+pub mod system_transaction;
 pub mod timing;
 pub mod token_program;
 pub mod transaction;
