+++ src/bin/bench-tps.rs
@@ -16,13 +16,13 @@ use solana::logger;
 use solana::ncp::Ncp;
 use solana::service::Service;
 use solana::signature::GenKeys;
-use solana::system_transaction::SystemTransaction;
 use solana::thin_client::{poll_gossip_for_leader, ThinClient};
 use solana::window::default_window;
 use solana_drone::drone::{request_airdrop_transaction, DRONE_PORT};
 use solana_metrics::influxdb;
 use solana_sdk::hash::Hash;
 use solana_sdk::signature::{read_keypair, Keypair, KeypairUtil};
+use solana_sdk::system_transaction::SystemTransaction;
 use solana_sdk::timing::timestamp;
 use solana_sdk::timing::{duration_as_ms, duration_as_s};
 use solana_sdk::transaction::Transaction;
