+++ src/rpc_pubsub.rs
@@ -253,7 +253,7 @@ mod tests {
     use solana_sdk::signature::{Keypair, KeypairUtil};
     use solana_sdk::transaction::Transaction;
     use std::net::{IpAddr, Ipv4Addr};
-    use system_transaction::SystemTransaction;
+    use solana_sdk::system_transaction::SystemTransaction;
     use tokio::prelude::{Async, Stream};
 
     #[test]
