+++ src/test_tx.rs
@@ -1,0 +1,11 @@
+use solana_sdk::hash::Hash;
+use solana_sdk::signature::{Keypair, KeypairUtil};
+use solana_sdk::system_transaction::*;
+use solana_sdk::transaction::Transaction;
+
+pub fn test_tx() -> Transaction {
+    let keypair1 = Keypair::new();
+    let pubkey1 = keypair1.pubkey();
+    let zero = Hash::default();
+    Transaction::system_new(&keypair1, pubkey1, 42, zero)
+}
