+++ tests/multinode.rs
@@ -24,12 +24,12 @@ use solana::packet::SharedBlob;
 use solana::poh_service::NUM_TICKS_PER_SECOND;
 use solana::result;
 use solana::service::Service;
-use solana::system_transaction::SystemTransaction;
 use solana::thin_client::{retry_get_balance, ThinClient};
 use solana::window::default_window;
 use solana_sdk::hash::Hash;
 use solana_sdk::pubkey::Pubkey;
 use solana_sdk::signature::{Keypair, KeypairUtil};
+use solana_sdk::system_transaction::SystemTransaction;
 use solana_sdk::timing::{duration_as_ms, duration_as_s};
 use solana_sdk::transaction::Transaction;
 use std::collections::{HashSet, VecDeque};
